[refactor] DTO 삭제 및 Tickets 불변화
