+++ src/main/java/domain/LottoResultMachine.java
@@ -1,9 +1,8 @@
 package domain;
 
-import domain.numberscontainer.Ticket;
+import domain.numberscontainer.Tickets;
 import domain.numberscontainer.WinningNumbers;
 
-import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
