--- src/main/java/domain/numberscontainer/LottoNumbersDto.java
@@ -1,16 +1,0 @@
-package domain.numberscontainer;
-
-import java.util.List;
-
-public class LottoNumbersDto {
-/*
-    private final List<String> lottoNumbers; //"1, 2, 3, 4, 5, 6"
-
-    public LottoNumbersDto(List<String> lottoNumbers) {
-        this.lottoNumbers = lottoNumbers;
-    }
-
-    public List<String> getLottoNumbers() {
-        return lottoNumbers;
-    }*/
-}
