+++ src/main/java/domain/numberscontainer/Ticket.java
@@ -6,10 +6,13 @@ import java.util.stream.Collectors;
 public class Ticket extends LottoNumbers {
 
     private static final String TICKET_DELIMITER = ", ";
-    private static final String STARTING_MARK = "[";
-    private static final String ENDING_MARK = "]";
+    private static final String START_MARK = "[";
+    private static final String END_MARK = "]";
+
+//    private final LottoNumbers lottoNumbers;
 
     public Ticket(String lottoNumbers) {
+//        this.lottoNumbers = lottoNumbers;
         super(lottoNumbers);
     }
 
@@ -25,6 +28,6 @@ public class Ticket extends LottoNumbers {
     public String toString() {
         return this.lottoNumbers.stream()
                 .map(lottoNumber -> Integer.toString(lottoNumber.getValue()))
-                .collect(Collectors.joining(TICKET_DELIMITER, STARTING_MARK, ENDING_MARK));
+                .collect(Collectors.joining(TICKET_DELIMITER, START_MARK, END_MARK));
     }
 }
