+++ src/test/java/domain/TicketsTest.java
@@ -1,5 +1,6 @@
 package domain;
 
+import domain.numberscontainer.Tickets;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
